feat: Implement deleteNode method in Solution class
This commit adds the implementation of the deleteNode method in the Solution class. The method is used to delete a given node in a singly-linked list. The node to be deleted is provided as an argument to the method. The implementation updates the value of the given node to the value of its next node, and updates the next pointer of the given node to skip the next node. This effectively removes the given node from the linked list.